# fix: Auto-detect Pascal GPUs and use FP32 for optimal performance