--- a/benchmark_results/implementation_comparison_20250630_132846.json
+++ b/benchmark_results/implementation_comparison_20250630_132846.json
@@ -1,0 +1,229 @@
+{
+  "config": {
+    "device": "cuda",
+    "dtype": "torch.float32",
+    "benchmark_config": {
+      "batch_sizes": [
+        2
+      ],
+      "seq_lengths": [
+        4096,
+        8192
+      ],
+      "num_heads_list": [
+        8
+      ],
+      "head_dim": 64,
+      "segment_lengths": [
+        [
+          512,
+          1024
+        ],
+        [
+          1024,
+          2048
+        ],
+        [
+          2048,
+          4096
+        ]
+      ],
+      "dilation_rates": [
+        [
+          1,
+          2
+        ],
+        [
+          1,
+          2
+        ],
+        [
+          1,
+          2
+        ]
+      ],
+      "warmup_steps": 3,
+      "benchmark_steps": 10,
+      "device": "cuda",
+      "use_fp16": true,
+      "use_memory_pool": true,
+      "memory_pool_size": null,
+      "use_pattern_cache": true,
+      "pattern_cache_size": 100,
+      "save_plots": true,
+      "save_csv": true,
+      "verbose": true
+    }
+  },
+  "results": [
+    {
+      "implementation": "standard",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 20.951616764068604,
+      "memory_mb": 173.0390625,
+      "throughput": 390996.07883478637,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "standard",
+      "batch_size": 2,
+      "seq_length": 8192,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 404.9520858563483,
+      "memory_mb": 332.0390625,
+      "throughput": 40459.10756417741,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "improved",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 5.435887631028891,
+      "memory_mb": 141.0390625,
+      "throughput": 1507021.5861782706,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "improved",
+      "batch_size": 2,
+      "seq_length": 8192,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 8.456532564014196,
+      "memory_mb": 268.0390625,
+      "throughput": 1937437.108646661,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "multihead_standard",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 134.65193305164576,
+      "memory_mb": 229.1640625,
+      "throughput": 60838.33937131789,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "multihead_standard",
+      "batch_size": 2,
+      "seq_length": 8192,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 271.9157881103456,
+      "memory_mb": 436.1640625,
+      "throughput": 60253.94889299786,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "multihead_improved",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 65.30570518225431,
+      "memory_mb": 341.1640625,
+      "throughput": 125440.80149104695,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "multihead_improved",
+      "batch_size": 2,
+      "seq_length": 8192,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 245.03612034022808,
+      "memory_mb": 692.1640625,
+      "throughput": 66863.61168815079,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "ring_v2_collective",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 4.248782247304916,
+      "memory_mb": 149.1640625,
+      "throughput": 1928081.8651500302,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "ring_v2_collective",
+      "batch_size": 2,
+      "seq_length": 8192,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 7.707538455724716,
+      "memory_mb": 308.1640625,
+      "throughput": 2125711.093641175,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "block_sparse",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 21.450721006840467,
+      "memory_mb": 137.666015625,
+      "throughput": 381898.5850120206,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "block_sparse",
+      "batch_size": 2,
+      "seq_length": 8192,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 111.45660346373916,
+      "memory_mb": 248.6669921875,
+      "throughput": 146998.91698503357,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "block_sparse_multihead",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 28.29658230766654,
+      "memory_mb": 198.1640625,
+      "throughput": 289504.92716502014,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "block_sparse_multihead",
+      "batch_size": 2,
+      "seq_length": 8192,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 150.31461082398891,
+      "memory_mb": 373.1640625,
+      "throughput": 108998.05355039549,
+      "success": true,
+      "error": null
+    }
+  ],
+  "timestamp": "20250630_132846"
+}