--- a/benchmark_results/implementation_comparison_20250630_140550.json
+++ b/benchmark_results/implementation_comparison_20250630_140550.json
@@ -1,0 +1,144 @@
+{
+  "config": {
+    "device": "cuda",
+    "dtype": "torch.float32",
+    "benchmark_config": {
+      "batch_sizes": [
+        2
+      ],
+      "seq_lengths": [
+        4096
+      ],
+      "num_heads_list": [
+        8
+      ],
+      "head_dim": 64,
+      "segment_lengths": [
+        [
+          512,
+          1024
+        ],
+        [
+          1024,
+          2048
+        ],
+        [
+          2048,
+          4096
+        ]
+      ],
+      "dilation_rates": [
+        [
+          1,
+          2
+        ],
+        [
+          1,
+          2
+        ],
+        [
+          1,
+          2
+        ]
+      ],
+      "warmup_steps": 3,
+      "benchmark_steps": 10,
+      "device": "cuda",
+      "use_fp16": true,
+      "use_memory_pool": true,
+      "memory_pool_size": null,
+      "use_pattern_cache": true,
+      "pattern_cache_size": 100,
+      "save_plots": true,
+      "save_csv": true,
+      "verbose": true
+    }
+  },
+  "results": [
+    {
+      "implementation": "standard",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 163.31249568611383,
+      "memory_mb": 173.0390625,
+      "throughput": 50161.50151636285,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "improved",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 38.52482456713915,
+      "memory_mb": 141.0390625,
+      "throughput": 212642.11043254434,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "multihead_standard",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 225.42250845581293,
+      "memory_mb": 229.1640625,
+      "throughput": 36340.647862171165,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "multihead_improved",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 136.19464542716742,
+      "memory_mb": 341.1640625,
+      "throughput": 60149.207586731616,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "ring_v2_collective",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 9.868623968213797,
+      "memory_mb": 149.1640625,
+      "throughput": 830105.5979421148,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "block_sparse",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 41.16707080975175,
+      "memory_mb": 137.666015625,
+      "throughput": 198993.99784498295,
+      "success": true,
+      "error": null
+    },
+    {
+      "implementation": "block_sparse_multihead",
+      "batch_size": 2,
+      "seq_length": 4096,
+      "num_heads": 8,
+      "head_dim": 64,
+      "time_ms": 152.74699544534087,
+      "memory_mb": 198.1640625,
+      "throughput": 53631.169478102325,
+      "success": true,
+      "error": null
+    }
+  ],
+  "timestamp": "20250630_140550"
+}