## Summary

This PR fixes a critical performance issue affecting all Pascal GPU users (GTX 1060/1070/1080, P100). The library was defaulting to FP16 on all CUDA devices, which causes an **8x performance degradation** on Pascal GPUs.

## Problem

Pascal GPUs have very limited FP16 performance:
- No Tensor Core support
- Limited FP16 compute units  
- In our tests: **FP16 is 8x slower than FP32**

### Benchmark Results
On NVIDIA GeForce GTX 1080:
- **FP32**: 5.75 ms (712,795 tokens/s)
- **FP16**: 47.14 ms (86,894 tokens/s)
- **FP16 "Speedup"**: 0.12x (actually 8x SLOWER\!)

## Solution

This PR adds smart dtype selection based on GPU architecture:

1. **Auto-detects GPU compute capability**
   - Pascal and older (compute < 7.0): Uses FP32
   - Volta and newer (compute >= 7.0): Uses FP16

2. **Warns about suboptimal choices**
   - If user explicitly uses FP16 on Pascal, shows warning
   - Explains performance impact

3. **Maintains backward compatibility**
   - Respects explicit dtype choices
   - Falls back to FP32 if detection fails

## Changes

- Added `gpu_utils.py` with architecture detection functions
- Updated `RingDilatedAttentionV2Collective` to use smart dtype selection
- Added comprehensive tests for dtype selection logic
- Added documentation explaining the optimization

## Testing

- ✅ All existing tests pass
- ✅ New tests for GPU detection and dtype selection
- ✅ Verified on real Pascal GPU (GTX 1080)
- ✅ Verification script shows correct behavior

## Impact

This change provides **8x speedup** for all Pascal GPU users without any code changes needed on their part.

### Affected GPUs
- GeForce GTX 1060, 1070, 1080, 1080 Ti
- Tesla P100, P40, P4
- Quadro P-series
- Titan X (Pascal), Titan Xp

## Test Plan

1. Run on Pascal GPU - verify FP32 is auto-selected
2. Run on Volta+ GPU - verify FP16 is auto-selected  
3. Explicitly set dtype - verify it's respected with warning
4. Run benchmarks - verify performance improvement

🤖 Generated with [Claude Code](https://claude.ai/code)